+++ spec/dummy/app/models/user.rb
@@ -1,7 +1,9 @@
 class User < ApplicationRecord
   include Federails::Entity
 
-  acts_as_federails_actor username_field: :id, name_field: :email, user_count_method: :user_count
+  acts_as_federails_actor username_field:    :id,
+                          name_field:        :email,
+                          user_count_method: :user_count
 
   # Include default devise modules. Others available are:
   # :confirmable, :lockable, :timeoutable, :trackable and :omniauthable