+++ lib/federails/configuration.rb
@@ -49,13 +49,6 @@ module Federails
     mattr_accessor :remote_follow_url_method
     @@remote_follow_url_method = 'federails.new_client_following_url'
 
-    # Attribute in the user model to use as the user's name
-    # @deprecated Set name_field option on acts_as_federails_actor instead
-    #
-    # It only have sense if you have a separate username attribute
-    mattr_accessor :user_name_field
-    @@user_name_field = nil
-
     # Attribute in the user model to use as the username for local actors
     # @deprecated Set username_field option on acts_as_federails_actor instead
     mattr_accessor :user_username_field