+++ lib/federails.rb
@@ -28,7 +28,6 @@ module Federails
       :server_routes_path,
       :client_routes_path,
       :remote_follow_url_method,
-      :user_username_field, # @deprecated
       :base_client_controller,
     ].each { |key| Configuration.send :"#{key}=", config[key] if config.key?(key) }
   end