+++ app/controllers/federails/client/actors_controller.rb
@@ -9,6 +9,7 @@ module Federails
         authorize Federails::Actor, policy_class: Federails::Client::ActorPolicy
 
         @actors = policy_scope(Federails::Actor, policy_scope_class: Federails::Client::ActorPolicy::Scope).all
+        @actors = @actors.local if params[:local_only]
       end
 
       # GET /app/actors/1