+++ app/policies/federails/client/actor_policy.rb
@@ -7,7 +7,7 @@ module Federails
 
       class Scope < Scope
         def resolve
-          scope.local
+          scope
         end
       end
     end