+++ spec/routing/federails/server/actors_routing_spec.rb
@@ -5,15 +5,15 @@ module Federails
     RSpec.describe ActorsController, type: :routing do
       describe 'routing' do
         it 'routes to #show' do
-          expect(get: '/federation/actors/1').to route_to('federails/server/actors#show', id: '1')
+          expect(get: '/federation/actors/1').to route_to('federails/server/actors#show', format: :activitypub, id: '1')
         end
 
         it 'routes to #followers' do
-          expect(get: '/federation/actors/1/followers').to route_to('federails/server/actors#followers', id: '1')
+          expect(get: '/federation/actors/1/followers').to route_to('federails/server/actors#followers', format: :activitypub, id: '1')
         end
 
         it 'routes to #following' do
-          expect(get: '/federation/actors/1/following').to route_to('federails/server/actors#following', id: '1')
+          expect(get: '/federation/actors/1/following').to route_to('federails/server/actors#following', format: :activitypub, id: '1')
         end
       end
     end