+++ CHANGELOG.md
@@ -37,8 +37,8 @@ Please, keep them in this order when updating.
 - 
 ### Removed
 
-- As actors' subject is a polymorphic relation, these Federails configuration options were removed: `user_class` and 
-  `user_table` 
+- As actors' subject is a polymorphic relation, these Federails configuration options were removed: `user_class`, 
+  `user_table` and `user_profile_url_method`
 
 ## [0.2.0] 2024-10-13 - Sign messages, handle signed messages and remote following