+++ CHANGELOG.md
@@ -34,7 +34,8 @@ Please, keep them in this order when updating.
   `base_client_controller` option
 - New generator: `federails:copy_client_views`, that copies all the client views in `app/views/federails/client` for override
 - Added `auto_create_actors` option for `acts_as_federails_actor` method to disable automatic actor creation.
-- 
+- Added helper method `Federails.actor_entity?` to check if a given class/instance may have associated actors
+
 ### Removed
 
 - As actors' subject is a polymorphic relation, these Federails configuration options were removed: `user_class`,