+++ CHANGELOG.md
@@ -55,6 +55,7 @@ Please, keep them in this order when updating.
 
 - Client controllers: enforce authorization calls on controller actions
 - Server controllers: enforce authorization calls on controller actions
+- Mime types: Don't consider "application/json" as "application/ld+json"
 
 ## [0.2.0] 2024-10-13 - Sign messages, handle signed messages and remote following