+++ CHANGELOG.md
@@ -28,6 +28,8 @@ Please, keep them in this order when updating.
 
 ## [Unreleased]
 
+## [0.3.0] 2024-11-23
+
 ### Added
 
 - Base controller for client controllers can be specified to something different from `ActionController::Base` with the 
@@ -35,6 +37,9 @@ Please, keep them in this order when updating.
 - New generator: `federails:copy_client_views`, that copies all the client views in `app/views/federails/client` for override
 - Added `auto_create_actors` option for `acts_as_federails_actor` method to disable automatic actor creation.
 - Added helper method `Federails.actor_entity?` to check if a given class/instance may have associated actors
+- Dynamic dispatch of activities with `after_activity_received` (e.g.: `after_activity_received 'Create', 'Note', :create_note`)
+- Ability to add custom data to actor responses
+- Handle URI-only objects in dynamic dispatch
 
 ### Changed